Changes the folder name
