--- Charts/Change-X-axis-label-position/.NET-Standard/Change-X-axis-label-position/Change-X-axis-label-position.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
-    <RootNamespace>Change_X_axis_label_position</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Syncfusion.DocIO.Net.Core" Version="*" />
-  </ItemGroup>
-
-</Project>
